requirements.txt
alpaca-py==0.40.1
annotated-types==0.7.0
blinker==1.8.2
bcrypt==3.2.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
msgpack==1.1.1
numpy==1.24.4
pandas==2.0.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sseclient-py==1.8.0
typing_inspect==0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
websockets==13.0.1
Werkzeug==3.0.6
